+++ src/Symfony/Component/Security/Core/Tests/User/InMemoryUserProviderTest.php
@@ -13,8 +13,8 @@ namespace Symfony\Component\Security\Core\Tests\User;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
+use Symfony\Component\Security\Core\User\InMemoryUser;
 use Symfony\Component\Security\Core\User\InMemoryUserProvider;
-use Symfony\Component\Security\Core\User\User;
 
 class InMemoryUserProviderTest extends TestCase
 {
@@ -25,20 +25,17 @@ class InMemoryUserProviderTest extends TestCase
         $user = $provider->loadUserByUsername('fabien');
         $this->assertEquals('foo', $user->getPassword());
         $this->assertEquals(['ROLE_USER'], $user->getRoles());
-        $this->assertFalse($user->isEnabled());
     }
 
     public function testRefresh()
     {
-        $user = new User('fabien', 'bar');
+        $user = new InMemoryUser('fabien', 'bar');
 
         $provider = $this->createProvider();
 
         $refreshedUser = $provider->refreshUser($user);
         $this->assertEquals('foo', $refreshedUser->getPassword());
         $this->assertEquals(['ROLE_USER'], $refreshedUser->getRoles());
-        $this->assertFalse($refreshedUser->isEnabled());
-        $this->assertFalse($refreshedUser->isCredentialsNonExpired());
     }
 
     protected function createProvider(): InMemoryUserProvider
@@ -46,7 +43,6 @@ class InMemoryUserProviderTest extends TestCase
         return new InMemoryUserProvider([
             'fabien' => [
                 'password' => 'foo',
-                'enabled' => false,
                 'roles' => ['ROLE_USER'],
             ],
         ]);
@@ -55,7 +51,7 @@ class InMemoryUserProviderTest extends TestCase
     public function testCreateUser()
     {
         $provider = new InMemoryUserProvider();
-        $provider->createUser(new User('fabien', 'foo'));
+        $provider->createUser(new InMemoryUser('fabien', 'foo'));
 
         $user = $provider->loadUserByUsername('fabien');
         $this->assertEquals('foo', $user->getPassword());
@@ -65,8 +61,8 @@ class InMemoryUserProviderTest extends TestCase
     {
         $this->expectException(\LogicException::class);
         $provider = new InMemoryUserProvider();
-        $provider->createUser(new User('fabien', 'foo'));
-        $provider->createUser(new User('fabien', 'foo'));
+        $provider->createUser(new InMemoryUser('fabien', 'foo'));
+        $provider->createUser(new InMemoryUser('fabien', 'foo'));
     }
 
     public function testLoadUserByUsernameDoesNotExist()